FaceRecognition/face_recognizer/teste_webcam.py
import cv2

cap = cv2.VideoCapture(0, cv2.CAP_DSHOW)

while True:
    ret, frame = cap.read()
    if not ret:
        print("Falha ao capturar da webcam")
        break

    cv2.imshow("Webcam Teste", frame)

    if cv2.waitKey(1) & 0xFF == ord('q'):
        break

cap.release()
cv2.destroyAllWindows()
