FaceRecognition/face_recognizer/requirements.txt
dlib==19.24.0
face-recognition==1.3.0
numpy==1.26.4
Pillow==9.4.0
pip==25.1.1
cmake
opencv-python
setuptools